+++ lib/src/enum/listener_type.dart
@@ -1,16 +1,17 @@
 /// callback类型
 class ListenerType {
-  static final simpleMsgListener = 'simpleMsgListener';
-  static final connectListener = 'connectListener';
-  static final userListener = 'userListener';
-  static final groupListener = 'groupListener';
-  static final advancedMsgListener = 'advancedMsgListener';
-  static final conversationListener = 'conversationListener';
-  static final friendListener = 'friendListener';
-  static final signalingListener = 'signalingListener';
-  static final msgSendProgressListener = "msgSendProgressListener";
-  static final workMomentsListener = "workMomentsListener";
-  static final organizationListener = "organizationListener";
-  static final customBusinessListener = "customBusinessListener";
-  static final messageKvInfoListener = "messageKvInfoListener";
+  static const simpleMsgListener = 'simpleMsgListener';
+  static const connectListener = 'connectListener';
+  static const userListener = 'userListener';
+  static const groupListener = 'groupListener';
+  static const advancedMsgListener = 'advancedMsgListener';
+  static const conversationListener = 'conversationListener';
+  static const friendListener = 'friendListener';
+  static const signalingListener = 'signalingListener';
+  static const msgSendProgressListener = "msgSendProgressListener";
+  static const workMomentsListener = "workMomentsListener";
+  static const organizationListener = "organizationListener";
+  static const customBusinessListener = "customBusinessListener";
+  static const messageKvInfoListener = "messageKvInfoListener";
+  static const listenerForService = "listenerForService";
 }