+++ lib/src/logger.dart
@@ -1,0 +1,9 @@
+import 'dart:developer';
+
+/// print full log
+class Logger {
+  // Sample of abstract logging function
+  static void print(String text) {
+    log('** $text', name: 'flutter_openim_sdk');
+  }
+}